Add the pipeline scenarios, the config files, and the correctness tier
The kernels answer whether a computation is faster compiled. The scenarios answer
the question underneath it: how much of a real estimator's post-nuisance work any
of those kernels is. Both are needed, and the second decides whether a 9x kernel is
worth having -- 9x on 2% of a pipeline is 1.8%.

Two flavours have no retarget to time and the reason is on the record rather than
worked around: LTMLE refuses one because g_bounds enters every earlier node's
pseudo-outcome, and CTMLE's candidates are themselves learner fits. Both report the
share of their fit that is external instead, and say so in the row.

The correctness tier is a fast-tier test rather than a benchmark switch. It checks
what makes a timing mean anything -- that the compiled paths compute what numpy
computes -- plus two things a timing run cannot see: that a parallel kernel's answer
does not move with the thread count, and that the fused multiplier computes the
statistic it claims to, checked against signs reconstructed from the same counter
hash rather than against a quantile that would pass a dropped column.

+++ benchmarks/numba/implementations/numpy_reference.py
@@ -97,7 +97,7 @@ def resolve_plan(mode: Mode | str, cores: int, *, workers: int | None = None) ->
     if resolved.name == "task_parallel":
         return ThreadPlan(numba_threads=1, blas_threads=1, workers=cores)
     if resolved.name == "hybrid":
-        count = workers if workers else max(1, int(round(cores**0.5)))
+        count = workers if workers else max(1, round(cores**0.5))
         count = max(1, min(count, cores))
         return ThreadPlan(numba_threads=max(1, cores // count), blas_threads=1, workers=count)
     raise KeyError(f"unknown mode {resolved.name!r}")
